ResponseException stores the JSON sent to Druid.
-- `DruidAggregationQuery` does not have `toString` defined.
Furthermore, it would be non-trivial to define a `toString` that
serializes the query into the actual query sent to Druid (we would need
to inject a `ObjectMapper` from the `ObjectMapperSuite` into the
aggregation query). Therefore, the `ResponseException` now takes an
`ObjectWriter` that can be used to serialize the druid query that is
logged.